+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes to this project are documented in this file. More detail and links can b
 
 ## 1.17.0
 ### Added
+- AUTOTOOL-1373: ([GitHub Issue 128](https://github.com/F5Networks/f5-declarative-onboarding/issues/128)) Support Failover Multicast on the BIG-IPs default device.
 - AUTOTOOL-1923: ([GitHub Issue 164](https://github.com/F5Networks/f5-declarative-onboarding/issues/164)) Rudimentary DeviceCertificate validation
 - AUTOTOOL-1943: Allow variable expressions in some RemoteAuthRole fields
 ### Fixed